requirements.txt
transformers>=4.21.0
torch>=1.9.0
langdetect>=1.0.9
nltk>=3.8
pydub>=0.25.1
requests>=2.28.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0